+++ src/__internal__/focus-trap/focus-trap.component.tsx
@@ -13,10 +13,9 @@ import {
   onTabGuardFocus,
   trapFunction,
 } from "./focus-trap-utils";
-import {
-  ModalContext,
+import ModalContext, {
   ModalContextProps,
-} from "../../components/modal/modal.component";
+} from "../../components/modal/__internal__/modal.context";
 import usePrevious from "../../hooks/__internal__/usePrevious";
 import TopModalContext from "../../components/carbon-provider/top-modal-context";
