+++ src/__internal__/popover/popover.component.tsx
@@ -5,10 +5,9 @@ import { flip, Placement, Middleware } from "@floating-ui/dom";
 import useFloating from "../../hooks/__internal__/useFloating";
 import { StyledBackdrop, StyledPopoverContent } from "./popover.style";
 import CarbonScopedTokensProvider from "../../style/design-tokens/carbon-scoped-tokens-provider/carbon-scoped-tokens-provider.component";
-import {
-  ModalContext,
+import ModalContext, {
   ModalContextProps,
-} from "../../components/modal/modal.component";
+} from "../../components/modal/__internal__/modal.context";
 
 type CustomRefObject<T> = {
   current?: T | null;
