+++ src/components/form/form.component.tsx
@@ -2,7 +2,7 @@ import React, { useRef, useContext } from "react";
 import { SpaceProps, PaddingProps } from "styled-system";
 
 import { SidebarContext } from "../sidebar/sidebar.component";
-import { ModalContext } from "../modal/modal.component";
+import ModalContext from "../modal/__internal__/modal.context";
 import FormSummary from "./__internal__/form-summary.component";
 import {
   StyledForm,
