+++ src/components/modal/__internal__/modal.context.ts
@@ -1,0 +1,9 @@
+import React from "react";
+
+export interface ModalContextProps {
+  isInModal?: boolean;
+  isAnimationComplete?: boolean;
+  triggerRefocusFlag?: boolean;
+}
+
+export default React.createContext<ModalContextProps>({});
